Refactor solver module by modularizing components into `versions` and `runtimes`.
